refactor: Add UI separator component
This code change adds a new UI separator component to the project. The component is implemented using the @radix-ui/react-separator dependency. The separator is used to visually separate different sections or elements in the user interface. This addition enhances the visual presentation and improves the user experience of the application.